feat: create dedicated core struct for backend
